libFLAC/bitwriter.c: Add sanity check to prevent excessive allocation
When fuzzing the encoder it is possible to cause the encoder to
allocate huge amounts of memory. Sanity check for the number of
bits to grow the bitwrite capacity and returning false (indicating
memory allocation failed) prevents this and seems to have no effect
in the encoding of non-fuzzing inputs.

Credit: Oss-Fuzz
Issue: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=17273
Testcase: fuzzer_encoder-5640245298593792